requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
crispy-bootstrap5==2025.4
cryptography==44.0.2
decorator==5.2.1
distro==1.9.0
Django==5.2
django-allauth==65.7.0
django-appconf==1.1.0
django-bootstrap5==25.1
django-ckeditor==6.7.2
django-crispy-forms==2.3
django-filter==25.1
django-formtools==2.5.1
django-js-asset==3.1.2
django-unfold==0.55.2
django-unicorn==0.62.0
djangorestframework==3.16.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
Markdown==3.8
oauthlib==3.2.2
openai==1.76.0
orjson==3.10.16
pillow==11.1.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
shortuuid==1.0.13
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
